# microsvc/backend/accounts-service/openapi.yml
openapi: 3.0.3
info:
  title: Accounts Service
  description: API for managing bank accounts.
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local development server
paths:
  /accounts:
    get:
      summary: Get all accounts for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
    post:
      summary: Create a new account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
  /accounts/{accountId}:
    get:
      summary: Get account by ID
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
  /accounts/{accountId}/balance:
    get:
      summary: Get account balance
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          description: Account not found
    put:
      summary: Update account balance
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - balance
              properties:
                balance:
                  type: number
      responses:
        '200':
          description: Balance updated successfully
        '404':
          description: Account not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        accountNumber:
          type: string
        balance:
          type: number
        accountType:
          type: string
    AccountCreateRequest:
      type: object
      required:
        - accountType
      properties:
        accountType:
          type: string
          enum: [CHECKING, SAVINGS]
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
        accountNumber:
          type: string